backend/requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt
gunicorn
psycopg2-binary
django-cors-headers
Pillow>=10.0.0
whitenoise
django-filter
celery
redis
django-celery-beat